// nnc-web/web/src/interfaces/core-response/ISession.ts
/* Copyright 2024 Sony Group Corporation. */
/**
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
*/

export interface IUrls {
  signin: string;
  account_settings: string;
}

export interface ICreateSession {
  urls: {
    top: IUrls;
    download: IUrls;
    dashboard: IUrls;
    abci: IUrls;
    webview: IUrls;
  };
  u: string;
  id_provider: 'google' | 'sony';
  auth_flow: 'sign-in' | 'sign-up';
}